==> Script/03_02_ML_for_candidate_gene.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#1. loading package and data
##1.1. 
```{r}
# devtools::install_local("../Data/MRS_1.0.0.tar.gz")
```
##1.2. 
```{r,message=F,warning=FALSE}
library(MRS)
library(magrittr)
library(data.table)
library(caret)
library(e1071)
library(pROC)
library(PRROC)
library(parallel)
library(doParallel)
library(tidyr)
library(dplyr)
library(ggpubr)
library(ggplot2)
library(RColorBrewer)
library(egg)
library(ComplexHeatmap)
library(Seurat)
```
##1.3. 
```{r}
Feature_ori <- readRDS("../Data/03_GeneInfo_forML.Rds")%>%as.data.table()
```

#2. 
##2.1. 
```{r}
Feature <- Feature_ori[, c(7:28)]
Feature <- Impute_with_label(Feature,method = "median")%>%as.data.table()
```
##2.2.
###2.2.1. 
```{r}
p <- Feature[Feature$Label == 1, ]
p <- p[, -1]
u <- Feature[Feature$Label != 1, ]
u <- u[, -1]
Parameter_select <- Tune_spy_pu(p_data = p, u_data = u, pu_repeats = 10, cv_folds = 3, seed = 123)
```
###2.2.2. 
```{r}
df <- Parameter_select
df %<>% mutate(
  "AUC±SD" = sprintf("%.3f±%.3f", AUC, AUC_SD),
  "AUPRC±SD" = sprintf("%.3f±%.3f", AUPRC, AUPRC_SD),
  "F1±SD" = sprintf("%.3f±%.3f", F1, F1_SD),
  "MCC±SD" = sprintf("%.3f±%.3f", MCC, MCC_SD)
)
display_df <- df[, c("spy_ratio","threshold","AUC±SD","AUPRC±SD","F1±SD","MCC±SD")]
p_table <- tableGrob(display_df, rows = NULL)
grid.draw(p_table)
```
##2.4.
```{r}
NewData <- Identify_reliable_negatives(Feature, spy_ratio =  0.3, threshold_quantile = 0.05,seed = 123)
Feature%<>%as.data.frame()
Feature <- Feature[,-which(colnames(Feature)=="Label")]
Feature$Name <- Feature_ori$ID
Allres <- merge(NewData,Feature)
Allres%<>%as.data.frame()
Allres <- Allres[ , -which(colnames(Allres)=="Name")]
anyNA(Allres)
saveRDS(Allres,"../Analysis/03_02_PosNegData.Rds")
# Allres <- readRDS("../Analysis/03_02_PosNegData.Rds")
```

#3.
##3.1. 
```{r}
prepared_data <- Prepare_classification_data(Allres, test_size = 0.3, seed = 123)
```
##3.2.
```{r}
Multi_models <- Train_multiple_models(prepared_data = prepared_data , n_folds = 10,n_repeats = 1, parallel = TRUE, n_cores=10, seed = 123)
```
##3.3.
###3.3.1. 
```{r}
Model_perf_Train <- Evaluate_train_performance(trained_models = Multi_models, prepared_data = prepared_data, seed = 123)
```
###3.3.2.
```{r}
per_fold_metrics <- Model_perf_Train$per_fold_metrics
#
fold_data_list <- lapply(names(per_fold_metrics), function(model_name) {
  df <- per_fold_metrics[[model_name]]
  df$Model <- model_name
  df <- as.data.table(df)
  
  #
  df[, AUC := as.numeric(AUC)]
  df[, AUPRC := as.numeric(AUPRC)]
  df[, F1 := as.numeric(F1)]
  df[, MCC := as.numeric(MCC)]
  
  return(df)
})

#
fold_data <- rbindlist(fold_data_list)

#
long_data <- pivot_longer(fold_data, cols = c(AUC, AUPRC, F1, MCC), 
                          names_to = "Metric", values_to = "Value")

#
summary_stats <- long_data %>%
  group_by(Model, Metric) %>%
  summarise(mean_val = mean(Value, na.rm = TRUE),
            sd_val = sd(Value, na.rm = TRUE),
            .groups = 'drop')

#
auc_order <- summary_stats %>%
  filter(Metric == "AUC") %>%
  arrange(desc(mean_val)) %>%
  pull(Model)

long_data$Model <- factor(long_data$Model, levels = auc_order)
summary_stats$Model <- factor(summary_stats$Model, levels = auc_order)
```
###3.3.2. 
```{r,fig.width=8,fig.height=4.5}
set.seed(123)
ggplot() +
  geom_bar(data = summary_stats, aes(x = Model, y = mean_val, fill = Metric),
           stat = "identity", position = position_dodge(width = 0.8), width = 0.7) +
  
  geom_errorbar(data = summary_stats, 
                aes(x = Model, ymin = mean_val - sd_val, ymax = mean_val + sd_val, group = Metric),
                position = position_dodge(width = 0.8), width = 0.3, color = "black") +
  
  geom_jitter(data = long_data, 
              aes(x = Model, y = Value, color = Metric),
              position = position_jitterdodge(jitter.width = 0.15, dodge.width = 0.8),
              size = 1.5, alpha = 0.8) +
  
  labs(title = "Model Performance on Training Data (10-fold CV)", y = "Metric Value", x = "Model") +
  theme_classic2(base_size = 12) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        legend.position = "top") +
  scale_fill_brewer(palette = "Set3",direction = -1) +
  scale_color_brewer(palette = "Dark2",direction = -1)+
  coord_cartesian(ylim = c(0.4, 1.05))
```
##3.4. 
###3.4.1. 
```{r}
Model_perf_Test <- Evaluate_test_performance(trained_models = Multi_models, prepared_data = prepared_data, seed = 123)
```
###3.4.2. 
```{r}
test_summary <- Model_perf_Test$test_results

#
test_summary$AUC_num <- as.numeric(test_summary$AUC)
test_summary$AUPRC_num <- as.numeric(test_summary$AUPRC)
test_summary$F1_num <- as.numeric(test_summary$F1)
test_summary$MCC_num <- as.numeric(test_summary$MCC)

#
long_test_data <- pivot_longer(test_summary, cols = c(AUC_num, AUPRC_num, F1_num, MCC_num), 
                          names_to = "Metric", values_to = "Value")

#
long_test_data$Metric <- recode(long_test_data$Metric,
                                "AUC_num" = "AUC",
                                "AUPRC_num" = "AUPRC",
                                "F1_num" = "F1",
                                "MCC_num" = "MCC")

#
auc_order <- test_summary %>%
  arrange(desc(AUC_num)) %>%
  pull(Model)

#
long_test_data$Model <- factor(long_test_data$Model, levels = auc_order)
```
###3.4.3. 
```{r,fig.width=8,fig.height=4.5}
ggplot() +
  geom_bar(data = long_test_data, aes(x = Model, y = Value, fill = Metric),
           stat = "identity", position = position_dodge(width = 0.8), width = 0.7) +
  labs(title = "Model Performance on Test Data", 
       y = "Metric Value", x = "Model") +
  theme_classic2(base_size = 12) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        legend.position = "top") +
  scale_fill_brewer(palette = "Set3",direction = -1) +
  coord_cartesian(ylim = c(0.4, 1))
```

#4. 
##4.1. 
```{r}
feature_selection <-
  Feature_selection_ablation(
    prepared_data = prepared_data,
    model_name = "XGBoost",
    n_folds = 10,
    n_repeats = 1,
    seed = 123,
    sampling = "up",
    parallel = TRUE,
    n_cores = 8,
    threshold_auc = 0.01, 
    threshold_auprc = 0.01, 
    threshold_f1 = 0.01, 
    threshold_mcc = 0.01, 
    center = TRUE, scale = TRUE
  )
```
##4.2. 
```{r,fig.width=12,fig.height=6}
feature_results <- feature_selection$feature_results
baseline_metrics <- feature_selection$baseline_metrics
#
fold_df <- rbindlist(lapply(names(feature_selection$per_fold_metrics), function(f) {
  df <- feature_selection$per_fold_metrics[[f]]
  df$Feature <- f
  return(df)
}))
fold_long <- melt(fold_df, id.vars = c("Feature", "Fold"), variable.name = "Metric", value.name = "Value")
fold_long <- merge(fold_long, feature_results[, c("Feature", "Retained", "Delta_MCC")], by = "Feature", all.x = TRUE)
fold_long$Feature <- factor(fold_long$Feature, 
                            levels = feature_results[order(-feature_results$Retained, -feature_results$Delta_MCC), "Feature"])
summary_df <- fold_long[, .(
  Mean = mean(Value, na.rm = TRUE),
  SD = sd(Value, na.rm = TRUE)
), by = .(Feature, Metric)]
summary_df$Baseline <- baseline_metrics[as.character(summary_df$Metric)]

#
ggplot(summary_df, aes(x = Feature, y = Mean)) +
  geom_bar(stat = "identity", aes(fill = Feature), width = 0.6, alpha = 0.6) +
  geom_errorbar(aes(ymin = Mean - SD, ymax = Mean + SD), width = 0.3, linewidth = 0.5) +
  geom_point(data = fold_long, aes(x = Feature, y = Value),color="black",
             position = position_jitter(width = 0.15), size = 1.2, alpha = 0.6, inherit.aes = FALSE) +
  geom_hline(aes(yintercept = Baseline), linetype = "dashed", color = "red", linewidth = 0.6) +
  facet_wrap(~ Metric, scales = "free_y", nrow = 4) +
  theme_bw(base_size = 13) +
  theme(
    axis.text.x = element_text(angle = 45, hjust = 1),
    strip.text = element_text(face = "bold"),
    strip.background = element_rect(fill = "#f7f7f7"),
    legend.position = "none",
    panel.grid = element_blank(),
    strip.background.x =  element_rect(fill = "white")
  ) +
  labs(
    title = "Feature Ablation: Cross-Validation Results (with SD)",
    x = "Feature", y = "Metric Value"
  ) +
  coord_cartesian(ylim = c(0.6, 1.1))
```

#5.
##5.1. 
```{r}
result <-
  Tune_model_eval(
    model_name = "XGBoost",
    prepared_data = feature_selection,
    n_folds = 10, n_repeats = 1,
    parallel = TRUE,n_cores = 15,
    tune_length = 500,search = "random", # "grid" "random"
    custom_grid = NULL,
    center = TRUE, scale = TRUE,
    sampling = "up",
    seed = 123
  )
```
##5.2.
```{r,fig.height=4,fig.width=8}
train_roc_list <- result$train_roc_list
test_roc <- result$test_roc
Plot_roc_curves(train_roc_list = train_roc_list,test_roc = test_roc)
```
##5.3. 
```{r,fig.height=4,fig.width=8}
train_pr_list <- result$train_pr_list
test_pr <- result$test_pr
Plot_pr_curves(train_pr_list = train_pr_list,test_pr = test_pr)
```
##5.4.
```{r,fig.width=4.5,fig.height=4.3}
#
Data <- result$train_metrics[,c("Fold","MCC","F1")]
Data%<>%melt()
#
summary_df <- Data %>%
  dplyr::group_by(variable) %>%
  summarise(
    Mean = mean(value, na.rm = TRUE),
    SD = sd(value, na.rm = TRUE)
  )
summary_df$variable%<>%as.character()
summary_df$variable <- factor(summary_df$variable,levels = c("F1","MCC"))
#
Data <- merge(Data, summary_df, by = "variable")

#
set.seed(123)
ggplot(summary_df, aes(x = variable, y = Mean, fill = variable)) +
  geom_col(width = 0.6, alpha = 0.8) +
  geom_errorbar(aes(ymin = Mean - SD, ymax = Mean + SD), width = 0.3, linewidth = 0.6) +
  geom_jitter(data = Data, aes(x = variable, y = value,color=variable), 
              width = 0.05, size = 2, alpha = 0.6, inherit.aes = FALSE) +
  theme_bw(base_size = 13) +
  theme(
    legend.position = "none",panel.grid = element_blank(),
    axis.text.x = element_text(angle = 0, hjust = 0.5)
  ) +
  labs(
    title = "Cross-Validation Performance per Metric",
    x = "Metric", y = "Score"
  ) +
  coord_cartesian(ylim = c(0.7, 1.02))+
  scale_fill_manual(values = c("#FFFFB3","#8DD3C7"))+
  scale_color_manual(values = c("#1B9E77","#D95F02"))
```
##5.5.
```{r,fig.width=4.5,fig.height=4.3}
Data <- result$final_test_metrics[,c("MCC","F1")]
Data%<>%melt()
Data$variable%<>%as.character()
Data$variable <- factor(Data$variable,levels = c("F1","MCC"))
#
set.seed(123)
ggplot(Data, aes(x = variable, y = value, fill = variable)) +
  geom_bar(stat = "identity", width = .5) +
  theme_bw(base_size = 13) +
  theme(
    legend.position = "none",panel.grid = element_blank(),
    axis.text.x = element_text(angle = 0, hjust = 0.5)
  ) +
  labs(
    title = "Cross-Validation Performance per Metric",
    x = "Metric", y = "Score"
  ) +
  coord_cartesian(ylim = c(0.7, 0.9))+
  scale_fill_manual(values = c("#FFFFB3","#8DD3C7"))
```
##5.6.
```{r,fig.width=5,fig.height=4}
#
Feature_Importance <- result$feature_importance %>% as.data.table()
setorder(Feature_Importance, Overall)
Feature_Importance$Feature <- factor(Feature_Importance$Feature, levels = Feature_Importance$Feature)

#
segment_color <- "#999999"
point_color <- "#444444"

#
ggplot(Feature_Importance, aes(x = Overall, y = Feature)) +
  geom_segment(aes(x = 0, xend = Overall, y = Feature, yend = Feature), 
               color = segment_color, size = 1) +
  geom_point(color = point_color, size = 3) +
  labs(
    x = "Scaled Importance Score",
    y = "Feature"
  ) +
  theme_bw(base_size = 13) +
  theme(
    axis.text.y = element_text(size = 10),
    axis.title = element_text(size = 12),
    plot.title = element_text(face = "bold", size = 14, hjust = 0.5),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank()
  ) +
  scale_x_continuous(expand = expansion(mult = c(0, 0.05)))
```

#6. 用模型进行预测 
##6.1. 加载所有基因 (中位数填充)
```{r}
Feature_ori <- readRDS("../Data/03_GeneInfo_forML.Rds")%>%as.data.table()
Feature <- Feature_ori[, c(7:28)]
Feature <- Impute_with_label(Feature)
Feature$Name <- Feature_ori$ID
Feature%<>%as.data.frame()
Feature <- Feature[,-1]
```
##6.2. 预测并赋予概率打分
```{r}
Predict_gene <- Feature[,feature_selection$selected_features]
Predict_classes <- predict(result$best_model, newdata = Predict_gene)
Predict_prop <- predict(result$best_model, newdata = Predict_gene, type = "prob")
Feature$Prop <- Predict_prop$pos
Feature%<>%as.data.table()
setkey(Feature,Name)
```
##6.3. 筛选Top10
```{r}
#
Pb_gene_ori <- readRDS("../Analysis/02_01_pb_mfuzz_gene.Rds")
Pf_gene_ori <- readRDS("../Analysis/02_01_pf_mfuzz_gene.Rds")
#
Gene <- intersect(Pf_gene_ori$Pf_gene,Pb_gene_ori$Pf_gene)
Gene <- gsub("\\..*$","",Gene)
Gene <- Feature[Gene]
setorder(Gene,-Prop)
Gene <- head(Gene,10)
```

#7.
##7.1. 
```{r}
ID_transform <- readRDS("/mnt/raid61/Personal_data/yangqingxin/data/PlasDB_Orthology/Science_genealign_inDiiffstrain.Rds")%>%as.data.table()
ID_transform$falciparum <- gsub("\\..*$","",ID_transform$falciparum)
setkey(ID_transform,berghei)
```
##7.2.
```{r}
pb_sce.big <- readRDS("../Analysis/01_pb_IDE_sc.Rds")
pf_sce.big <-  readRDS("../Analysis/01_pf_IDE_sc.Rds")
#
Gene_pf <- Gene$Name
setkey(ID_transform,falciparum)
Gene_pb <- ID_transform[Gene_pf]$berghei
Gene_pf <- gsub("_","-",Gene_pf)
Gene_pb <- gsub("_","-",Gene_pb)
```
##7.3.
```{r}
##
pb_express<- AverageExpression(object = pb_sce.big,features = Gene_pb,group.by = "Cluster")
pb_express <- pb_express$RNA
pb_express%<>%t()%>%scale()%>%t()

##
pf_express<- AverageExpression(object = pf_sce.big,features = Gene_pf,group.by = "Scmap")
pf_express <- pf_express$RNA
pf_express%<>%t()%>%scale()%>%t()

stopifnot(colnames(pb_express)==colnames(pf_express))
rownames(pf_express) <- gsub("-","_",rownames(pf_express))
rownames(pb_express) <- rownames(pf_express) 

#
colnames(pb_express) <- c("ER","MR","ET","MT","LT","ES","MS","LS")
colnames(pf_express) <- c("ER","MR","ET","MT","LT","ES","MS","LS")
```
##7.3. 
```{r}
Gene$Signal_Exported <- c(0,0,0,0,0,0,0,1,0,0)
#
Gene$Signal_peptide <- c(0,0,0,0,0,0,0,1,0,0)
#
pb_express <- pb_express[Gene$Name,]
pf_express <- pf_express[Gene$Name,]
```
##7.4. 
```{r}
p1 <- ComplexHeatmap::Heatmap(pf_express,
                              cluster_rows = FALSE,
                              cluster_columns = FALSE,
                              show_row_names = TRUE,
                              show_column_names = TRUE,
                              row_names_side = "left",
                              col = circlize::colorRamp2(c(-1, 0, 1), c("#2166AC", "#F7F7F7", "#B2182B")),
                              show_heatmap_legend = FALSE,
                              column_title = "P. falciparum",
                              column_names_side = "top",
                              column_names_rot = 45)
#
combined_annotation <- ComplexHeatmap::rowAnnotation(
  barplot = anno_barplot(Gene$Prop,
                         which = "row",
                         baseline = 0.9932,
                         extend = 0.15,
                         gp = gpar(fill = rep("black", length(Gene$Prop)))),
  bar2 = anno_barplot(c(0, 0, 0, 0, 0, 0, 0, 1, 0, 0)),
  bar3 = anno_barplot(c(0, 0, 0, 0, 0, 0, 0, 1, 0, 0))
)
#
p2 <- ComplexHeatmap::Heatmap(pb_express,
                              cluster_rows = FALSE,
                              cluster_columns = FALSE,
                              show_row_names = TRUE,
                              show_column_names = TRUE,
                              row_names_side = "left",
                              col = circlize::colorRamp2(c(-1, 0, 1), c("#2166AC", "#F7F7F7", "#B2182B")),
                              name = "Expression",
                              right_annotation = combined_annotation,
                              column_title = "P. berghei",
                              column_names_side = "top",
                              column_names_rot = 45)
```
##7.5.
```{r,fig.height=3,fig.width=7}
ht <- p1 + p2
draw(ht)
```
